Have Repsotory return Candidate instead of Version
